requirements.txt
asgiref==3.8.1
argon2-cffi==21.3.0
attrs==24.2.0
cffi==1.17.1
crispy-bootstrap5==2024.2
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fido2==1.1.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
psycopg2==2.9.9
pycparser==2.22
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
setuptools
selenium
webdriver-manager
beautifulsoup4
pdfplumber
requests
pymongo==4.10.1
undetected-chromedriver
PYPDF2